+++ src/quickjs_utils/sets.rs
@@ -1,0 +1,4 @@
+//! Set utils, these methods can be used to manage Set objects from rust
+//! see [MDN](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Sap) for more on Sets
+
+//todo