+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # 0.1.1 (Work in progress)
 
 * more precise timing for setTimeout and setInterval
+* quickjs_utils::maps utils for handling [Map](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Map) instances from rust
+* quickjs_utils::sets utils for handling [Set](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Set) instances from rust
+* Proxy.event_target and Proxy.static_event_target to allow a Proxy to be used as [EventTarget](https://developer.mozilla.org/en-US/docs/Web/API/EventTarget)
+* objects::construct_object util to create a new instance of a constructor
 
 # 0.1.0